+++ src/main/java/com/company/service/ILocationService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface ILocationService {
     List<Location> findAll();
-    Location findById(Long id) throws Exception;
+    Location findById(int id) throws Exception;
     Location findByCountryAndCityAndState(String country, String city, String state);
-    Location update(Long id, Location location) throws Exception;
+    Location update(int id, Location location) throws Exception;
     Location save(Location location) throws Exception;
-    void deleteById(Long id) throws Exception;
+    void deleteById(int id) throws Exception;
 }
